+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - change pitch wheel event slider's origin
     - chase option (under playback)
     - session not seen as modified when editing sequence
+    - fix Aftertouch editing
 
 ## 1.0.0